requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
langchain==0.0.340
langchain-openai==0.0.2
python-dotenv==1.0.0
pymongo==4.5.0
redis==5.0.1